+++ src/ng-vcl/password-input/password-input.component.ts
@@ -1,5 +1,6 @@
-import { Directive, Input, Output, ElementRef, HostListener, EventEmitter, Component, forwardRef, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
+import { Directive, Input, Output, ElementRef, HostListener, EventEmitter, Component, forwardRef, ChangeDetectionStrategy, ChangeDetectorRef, ContentChild, OnChanges, SimpleChanges, OnInit } from '@angular/core';
 import { ControlValueAccessor, FormsModule, NG_VALUE_ACCESSOR } from '@angular/forms';
+import { InputDirective } from './../input/index';
 
 export const CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR: any = {
   provide: NG_VALUE_ACCESSOR,
@@ -11,16 +12,14 @@ export const CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR: any = {
   templateUrl: 'password-input.component.html',
   selector: 'vcl-password-input',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  providers: [CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR],
   host: {
     '[class.vclInputGroup]': 'true',
     '[attr.tabindex]': '-1'
   }
 })
-export class PasswordInputComponent implements ControlValueAccessor {
-
-  @Input()
-  inputId: string | undefined;
+export class PasswordInputComponent implements OnInit, OnChanges  {
+  @ContentChild(InputDirective, { read: InputDirective })
+  input?: InputDirective;
 
   @Input()
   visibleIcon = 'fa:eye-slash';
@@ -34,54 +33,29 @@ export class PasswordInputComponent implements ControlValueAccessor {
   @Input()
   disabled = false;
 
-  @Input()
-  selectOnFocus = false;
-
-  @Input()
-  tabindex = 0;
-
-  @Input()
-  placeholder = '';
-
-  value = '';
-
   get buttonIcon() {
     return this.visible ? this.visibleIcon : this.invisibleIcon;
   }
 
-  constructor(private cdRef: ChangeDetectorRef) { }
+  ngOnInit(): void {
+    this.updateInput();
+  }
+
+  ngOnChanges(changes: SimpleChanges): void {
+    if (changes.visible || changes.disabled) {
+      this.updateInput();
+    }
+  }
+
+  updateInput() {
+    if (this.input) {
+      this.input.setType(this.visible ? 'text' : 'password');
+      this.input.setDisabled(this.disabled);
+    }
+  }
 
   toggle() {
     this.visible = !this.visible;
-    this.cdRef.markForCheck();
-  }
-
-  onBlur() {
-    this.onTouched();
-  }
-
-  onModelChange(value: any) {
-    this.value = String(value);
-    this.onChange(this.value);
-  }
-
-  /**
-   * Things needed for ControlValueAccessor-Interface.
-   */
-  private onChange: (_: any) => void = () => {};
-  private onTouched: () => any = () => {};
-  writeValue(value: any): void {
-    this.value = (value === null || value === undefined) ? '' : String(value);
-    this.cdRef.markForCheck();
-  }
-  registerOnChange(fn: any) {
-    this.onChange = fn;
-  }
-  registerOnTouched(fn: any) {
-    this.onTouched = fn;
-  }
-  setDisabledState(isDisabled: boolean) {
-    this.disabled = isDisabled;
-    this.cdRef.markForCheck();
+    this.updateInput();
   }
 }